+++ 05.sh
@@ -3,7 +3,7 @@
 # read http://docs.ceph.com/docs/master/install/manual-deployment/ for better understanding
 
 set -o nounset
-set -o errexit
+#set -o errexit
 set -o noclobber
 set -o noglob
 
@@ -36,15 +36,15 @@ MON_KEYRING="${CPWD}/${CLUSTER_NAME}.mon.keyring"
 OSD_UID=$(uuidgen)
 CLUSTER_FSID=$(grep fsid ${CLUSTER_CONF} | cut -d " " -f3)
 echo ${CLUSTER_FSID}
-OSD_ID=$(ceph --cluster ${CLUSTER_NAME} osd create) && echo ${OSD_ID}
-
+#OSD_ID=$(ceph --cluster ${CLUSTER_NAME} osd create) && echo ${OSD_ID}
+OSD_ID=$(ceph --cluster ${CLUSTER_NAME} osd create ${OSD_UID}) && echo ${OSD_ID} ${OSD_UID}
 
 # Create the default directories
 mkdir -p /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}
 
 ## prepare hdd
-#sgdisk -Z /dev/${CDISK}
-#${PARTPROBE} /dev/${CDISK}
+sgdisk -Z /dev/${CDISK}
+${PARTPROBE} /dev/${CDISK}
 #
 #sgdisk -n 1:0:100M /dev/${CDISK}
 #sgdisk -n 2 /dev/${CDISK}
@@ -59,12 +59,15 @@ mkdir -p /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}
 ## format and mount hdd
 #mkfs.xfs -f -i size=2048 -L "blueFS${OSD_ID}" /dev/${CDISK}1
 
-${CEPHDISK} zap /dev/${CDISK}
+${CEPHDISK} --zap-disk /dev/${CDISK}
+sleep 2
 ${PARTPROBE} /dev/${CDISK}
-${CEPHDISK} prepare --cluster ${CLUSTER_NAME} --bluestore --block.db ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.db --block.wal ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.wal /dev/${CDISK}
-
-#${PARTPROBE} /dev/${CDISK}
-#sleep 5
+sleep 2
+#${CEPHDISK} prepare --bluestore --block.db-file ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.db --block.wal-file ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.wal --osd-id ${OSD_ID} -osd-uuid ${OSD_UID} --journal-file ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.journal /dev/${CDISK}
+${CEPHDISK} prepare --bluestore --block.db ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.db --block.wal ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.wal --osd-id ${OSD_ID} --osd-uuid ${OSD_UID} --journal-file ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.journal /dev/${CDISK}
+#${CEPHDISK} prepare --bluestore --block.db ${JOURNALPOINT}/blockdb-for-${CDISK}.db --block.wal ${JOURNALPOINT}/wal-for-${CDISK}.wal /dev/${CDISK}
+sleep 2
+${PARTPROBE} /dev/${CDISK}
 
 # put it into fstab
 #echo "/dev/${CDISK}1	/var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}        xfs     rw,noexec,nodev,noatime   0 0" >> /etc/fstab
@@ -82,20 +85,28 @@ mount -a
 echo "[osd.${OSD_ID}]" >> ${CLUSTER_CONF}
 echo "host = ${LMON}" >> ${CLUSTER_CONF}
 #echo "osd data = /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}" >> ${CLUSTER_CONF}
-echo "bluestore block path = /dev/${CDISK}2" >> ${CLUSTER_CONF}
+echo "bluestore block path = /dev/disk/by-partuuid/$(blkid /dev/${CDISK}2 | grep PARTUUID | cut -d '"' -f2)" >> ${CLUSTER_CONF}
 echo "bluestore block db path = ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.db" >> ${CLUSTER_CONF}
 echo "bluestore block wal path = ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.wal" >> ${CLUSTER_CONF}
 echo "osd journal = ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.journal" >> ${CLUSTER_CONF}
 echo -n -e "\n" >> ${CLUSTER_CONF}
 
-echo "bluestore" > /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/type
+
+if [[ -f /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/type ]]; then
+	rm -f /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/type
+	echo "bluestore" > /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/type
+else
+	echo "bluestore" > /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/type
+fi
 
 
+#ln -sf /dev/disk/by-partuuid/$(blkid /dev/${CDISK}2 | grep PARTUUID | cut -d '"' -f2) /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/block
+#ln -sf ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.db /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/block.db
+#ln -sf ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.wal /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/block.wal
+#ln -sf ${JOURNALPOINT}/${CLUSTER_NAME}-${OSD_ID}.journal /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}/journal
 
+chown -R ceph. /var/lib/ceph/osd/${CLUSTER_NAME}-${OSD_ID}
 
-
-
-
-
+systemctl start ceph-osd@${OSD_ID}.service
 
 exit 0
